+++ tests/test_storage.py
@@ -147,13 +147,13 @@ def update_feed_last_updated(storage, feed, entry):
 
 def add_or_update_entry(storage, feed, entry):
     storage.add_or_update_entry(
-        EntryUpdateIntent(entry, entry.updated, datetime(2010, 1, 1), 0)
+        EntryUpdateIntent(entry, entry.updated, datetime(2010, 1, 1), 0, 0)
     )
 
 
 def add_or_update_entries(storage, feed, entry):
     storage.add_or_update_entries(
-        [EntryUpdateIntent(entry, entry.updated, datetime(2010, 1, 1), 0)]
+        [EntryUpdateIntent(entry, entry.updated, datetime(2010, 1, 1), 0, 0)]
     )
 
 
@@ -254,7 +254,7 @@ def check_errors_locked(db_path, pre_stuff, do_stuff, exc_type):
     entry = EntryData('one', 'entry', datetime(2010, 1, 2))
     storage.add_feed(feed.url, datetime(2010, 1, 1))
     storage.add_or_update_entry(
-        EntryUpdateIntent(entry, entry.updated, datetime(2010, 1, 1), 0)
+        EntryUpdateIntent(entry, entry.updated, datetime(2010, 1, 1), 0, 0)
     )
 
     in_transaction = threading.Event()
@@ -350,12 +350,12 @@ def check_iter_locked(db_path, pre_stuff, iter_stuff):
     entry = EntryData('one', 'entry', datetime(2010, 1, 1), title='entry')
     storage.add_feed(feed.url, datetime(2010, 1, 2))
     storage.add_or_update_entry(
-        EntryUpdateIntent(entry, entry.updated, datetime(2010, 1, 1), 0)
+        EntryUpdateIntent(entry, entry.updated, datetime(2010, 1, 1), 0, 0)
     )
     storage.add_feed('two', datetime(2010, 1, 1))
     storage.add_or_update_entry(
         EntryUpdateIntent(
-            entry._replace(feed_url='two'), entry.updated, datetime(2010, 1, 1), 0
+            entry._replace(feed_url='two'), entry.updated, datetime(2010, 1, 1), 0, 0
         )
     )
     storage.set_feed_metadata('two', '1', 1)
@@ -444,11 +444,12 @@ def test_get_entries_for_update(storage_cls):
             datetime(2010, 1, 2),
             datetime(2010, 1, 1),
             0,
+            0,
         )
     )
 
     assert list(storage.get_entries_for_update([('feed', 'one'), ('feed', 'two')])) == [
-        EntryForUpdate(datetime(2010, 1, 1), entry.hash),
+        EntryForUpdate(datetime(2010, 1, 1), entry.hash, 0),
         None,
     ]
 
@@ -467,6 +468,7 @@ def test_entry_remains_read_after_update(storage_with_two_entries):
             EntryData('feed', 'one', datetime(2010, 1, 1)),
             datetime(2010, 1, 2),
             datetime(2010, 1, 2),
+            0,
             0,
         )
     )
@@ -488,6 +490,7 @@ def storage_with_two_entries(storage):
             datetime(2010, 1, 2),
             datetime(2010, 1, 2),
             0,
+            0,
         )
     )
     storage.add_or_update_entry(
@@ -496,6 +499,7 @@ def storage_with_two_entries(storage):
             datetime(2010, 1, 2),
             datetime(2010, 1, 2),
             1,
+            0,
         )
     )
     return storage
@@ -548,6 +552,7 @@ def test_important_entry_remains_important_after_update(storage):
             EntryData('feed', 'one', datetime(2010, 1, 1)),
             datetime(2010, 1, 2),
             datetime(2010, 1, 2),
+            0,
             0,
         )
     )